make Decoder fields public
From experiments with an stm32f401re, the ARM specification may not be
followed in practise. Whe the decoder returns an error, it may thus be
necessary to manually (outside of the crate) inspect the incoming bytes
and eventually modify them to handle any possible target-specific
bitstreams which diverge from the spec. So instead of figuring out a
decent API for this intervention, we'll just make the internal state of
the decoder public.