refactor: simplify Decoder state
Instead of maintaining a state machine, the first byte (header) is now
instead decoded into a stub. This stub contains information on how many
more bytes are required to decode the whole packet. The incoming buffer
is then checked for lengh and drained appropriately.

This commit also removes public access to internal variables; the state
of the decoder is now always known.

Closes #9.