refactor: take Read instance, adopt an iter-based API
This commit is a complete rewrite of the decoder API. Instead of
manually pushing bytes to the decoder it now takes a std::io::Read
instance and handles I/O itself. This mimics the API offered by
rust-embedded/itm. Improving on this API is the implementation of
Iterator that yield different packet types. At present two iterators are
available: one that yield single packets and another one that yield a
set of packets, wrapped in an chrono::DateTime timestamp (with data
relation) based on the ITM timestamp clock and prescaler (user-supplied).

Some work still remains: the Buffer implementation must be reviewed and
possibly optimized; some better names should replace Decoder::{singles,
timestamps}; more documentation must be added; and the test bench for
Timestamps needs to be extended. The purpose of committing here is to
stop myself from amending an ever-growing single commit and prepare the
crate for comments now that it will replace rust-embedded/itm.

Because of this rewrite, and because of the plan to implement an
iterator that handles instrumentation/extention packet arbitration and
decoding, most functionality has been removed from the itm-decode
executable. It's purpose is to act as a thin wrapper around the library
for input files/pipes. It's (albeit non-conformant) instumentation
packet functionality will eventually be implemented in the library
itself.